requirements.txt
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.0
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.2.0
requests==2.28.1
Werkzeug==2.2.1
zipp==3.8.1
